Remove dry-run
