+++ README.md
@@ -150,8 +150,6 @@ Commands:
     predict      Use a trained model to make predictions.
     elmo         Create word vectors using a pretrained ELMo model.
     fine-tune    Continue training a model on a new dataset.
-    dry-run      Create a vocabulary, compute dataset statistics and other
-                 training utilities.
     make-vocab   Create a vocabulary, compute dataset statistics and other
                  training utilities.
     test-install
