+++ allennlp/commands/__init__.py
@@ -5,7 +5,6 @@ from typing import Any, Optional
 from overrides import overrides
 
 from allennlp import __version__
-from allennlp.commands.dry_run import DryRun
 from allennlp.commands.elmo import Elmo
 from allennlp.commands.evaluate import Evaluate
 from allennlp.commands.find_learning_rate import FindLearningRate
