+++ allennlp/tests/commands/train_test.py
@@ -9,14 +9,13 @@ from typing import Iterable
 import pytest
 import torch
 
-from allennlp.commands.train import Train, TrainModel, train_model, train_model_from_args
+from allennlp.commands.train import Train, train_model, train_model_from_args, TrainModel
 from allennlp.common import Params
 from allennlp.common.checks import ConfigurationError
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data import DatasetReader, Instance, Vocabulary
 from allennlp.models import load_archive, Model
 from allennlp.models.archival import CONFIG_NAME
-from allennlp.modules.token_embedders.embedding import _read_pretrained_embeddings_file
 
 SEQUENCE_TAGGING_DATA_PATH = str(AllenNlpTestCase.FIXTURES_ROOT / "data" / "sequence_tagging.tsv")
 
@@ -167,7 +166,7 @@ class TestTrain(AllenNlpTestCase):
         train_model(params, serialization_dir=serialization_dir)
 
         config_path = os.path.join(serialization_dir, CONFIG_NAME)
-        with open(config_path, "r") as config:
+        with open(config_path) as config:
             saved_config_as_dict = OrderedDict(json.load(config))
         assert params_as_dict == saved_config_as_dict
 
@@ -362,3 +361,152 @@ class TestTrainOnLazyDataset(AllenNlpTestCase):
         shutil.rmtree(serialization_dir, ignore_errors=True)
         with pytest.raises(Exception):
             train_model(params, serialization_dir=serialization_dir)
+
+
+class TestDryRun(AllenNlpTestCase):
+    def setUp(self):
+        super().setUp()
+
+        self.params = Params(
+            {
+                "model": {
+                    "type": "simple_tagger",
+                    "text_field_embedder": {
+                        "token_embedders": {"tokens": {"type": "embedding", "embedding_dim": 5}}
+                    },
+                    "encoder": {"type": "lstm", "input_size": 5, "hidden_size": 7, "num_layers": 2},
+                },
+                "dataset_reader": {"type": "sequence_tagging"},
+                "train_data_path": str(self.FIXTURES_ROOT / "data" / "sequence_tagging.tsv"),
+                "validation_data_path": str(self.FIXTURES_ROOT / "data" / "sequence_tagging.tsv"),
+                "iterator": {"type": "basic", "batch_size": 2},
+                "trainer": {"num_epochs": 2, "optimizer": "adam"},
+            }
+        )
+
+    def test_dry_run_doesnt_overwrite_vocab(self):
+        vocab_path = self.TEST_DIR / "vocabulary"
+        os.mkdir(vocab_path)
+        # Put something in the vocab directory
+        with open(vocab_path / "test.txt", "a+") as open_file:
+            open_file.write("test")
+        # It should raise error if vocab dir is non-empty
+        with pytest.raises(ConfigurationError):
+            train_model(self.params, self.TEST_DIR, dry_run=True)
+
+    def test_dry_run_without_vocabulary_key(self):
+        train_model(self.params, self.TEST_DIR, dry_run=True)
+
+    def test_dry_run_makes_vocab(self):
+        vocab_path = self.TEST_DIR / "vocabulary"
+
+        train_model(self.params, self.TEST_DIR, dry_run=True)
+
+        vocab_files = os.listdir(vocab_path)
+        assert set(vocab_files) == {"labels.txt", "non_padded_namespaces.txt", "tokens.txt"}
+
+        with open(vocab_path / "tokens.txt") as f:
+            tokens = [line.strip() for line in f]
+
+        tokens.sort()
+        assert tokens == [".", "@@UNKNOWN@@", "animals", "are", "birds", "cats", "dogs", "snakes"]
+
+        with open(vocab_path / "labels.txt") as f:
+            labels = [line.strip() for line in f]
+
+        labels.sort()
+        assert labels == ["N", "V"]
+
+    def test_dry_run_with_extension(self):
+        existing_serialization_dir = self.TEST_DIR / "existing"
+        extended_serialization_dir = self.TEST_DIR / "extended"
+        existing_vocab_path = existing_serialization_dir / "vocabulary"
+        extended_vocab_path = extended_serialization_dir / "vocabulary"
+
+        vocab = Vocabulary()
+        vocab.add_token_to_namespace("some_weird_token_1", namespace="tokens")
+        vocab.add_token_to_namespace("some_weird_token_2", namespace="tokens")
+        os.makedirs(existing_serialization_dir, exist_ok=True)
+        vocab.save_to_files(existing_vocab_path)
+
+        self.params["vocabulary"] = {}
+        self.params["vocabulary"]["type"] = "extend"
+        self.params["vocabulary"]["directory"] = str(existing_vocab_path)
+        self.params["vocabulary"]["min_count"] = {"tokens": 3}
+        train_model(self.params, extended_serialization_dir, dry_run=True)
+
+        vocab_files = os.listdir(extended_vocab_path)
+        assert set(vocab_files) == {"labels.txt", "non_padded_namespaces.txt", "tokens.txt"}
+
+        with open(extended_vocab_path / "tokens.txt") as f:
+            tokens = [line.strip() for line in f]
+
+        assert tokens[0] == "@@UNKNOWN@@"
+        assert tokens[1] == "some_weird_token_1"
+        assert tokens[2] == "some_weird_token_2"
+
+        tokens.sort()
+        assert tokens == [
+            ".",
+            "@@UNKNOWN@@",
+            "animals",
+            "are",
+            "some_weird_token_1",
+            "some_weird_token_2",
+        ]
+
+        with open(extended_vocab_path / "labels.txt") as f:
+            labels = [line.strip() for line in f]
+
+        labels.sort()
+        assert labels == ["N", "V"]
+
+    def test_dry_run_without_extension(self):
+        existing_serialization_dir = self.TEST_DIR / "existing"
+        extended_serialization_dir = self.TEST_DIR / "extended"
+        existing_vocab_path = existing_serialization_dir / "vocabulary"
+        extended_vocab_path = extended_serialization_dir / "vocabulary"
+
+        vocab = Vocabulary()
+        # if extend is False, its users responsibility to make sure that dataset instances
+        # will be indexible by provided vocabulary. At least @@UNKNOWN@@ should be present in
+        # namespace for which there could be OOV entries seen in dataset during indexing.
+        # For `tokens` ns, new words will be seen but `tokens` has @@UNKNOWN@@ token.
+        # but for 'labels' ns, there is no @@UNKNOWN@@ so required to add 'N', 'V' upfront.
+        vocab.add_token_to_namespace("some_weird_token_1", namespace="tokens")
+        vocab.add_token_to_namespace("some_weird_token_2", namespace="tokens")
+        vocab.add_token_to_namespace("N", namespace="labels")
+        vocab.add_token_to_namespace("V", namespace="labels")
+        os.makedirs(existing_serialization_dir, exist_ok=True)
+        vocab.save_to_files(existing_vocab_path)
+
+        self.params["vocabulary"] = {}
+        self.params["vocabulary"]["type"] = "from_files"
+        self.params["vocabulary"]["directory"] = str(existing_vocab_path)
+        train_model(self.params, extended_serialization_dir, dry_run=True)
+
+        with open(extended_vocab_path / "tokens.txt") as f:
+            tokens = [line.strip() for line in f]
+
+        assert tokens[0] == "@@UNKNOWN@@"
+        assert tokens[1] == "some_weird_token_1"
+        assert tokens[2] == "some_weird_token_2"
+        assert len(tokens) == 3
+
+    def test_make_vocab_args(self):
+        parser = argparse.ArgumentParser(description="Testing")
+        subparsers = parser.add_subparsers(title="Commands", metavar="")
+        Train().add_subparser(subparsers)
+        for serialization_arg in ["-s", "--serialization-dir"]:
+            raw_args = [
+                "train",
+                "path/to/params",
+                serialization_arg,
+                "serialization_dir",
+                "--dry-run",
+            ]
+            args = parser.parse_args(raw_args)
+            assert args.func == train_model_from_args
+            assert args.param_path == "path/to/params"
+            assert args.serialization_dir == "serialization_dir"
+            assert args.dry_run
