+++ training_config/event2mind.json
@@ -1,7 +1,7 @@
 {
   "dataset_reader": {
     "type": "event2mind",
-    # Uncomment this when generating the vocabularly with `dry-run`.
+    # Uncomment this when generating the vocabularly with `--dry-run`.
     #"dummy_instances_for_vocab_generation": true,
     "source_tokenizer": {
       "type": "spacy",
@@ -14,7 +14,7 @@
     }
   },
   "vocabulary": {
-    # Uncomment this when generating the vocabularly with `dry-run`.
+    # Uncomment this when generating the vocabularly with `--dry-run`.
     #"min_count": {"source_tokens": 2}
     # Uncomment this when training using an existing vocabularly.
     #"directory_path": "output_dir/vocabulary/"
